Basic Drivable Robot and Colliders
A basic framework for robot creation, input, and functionality was implement through a base class and an interface to layout the rules for how a robot class is formed. a game manager class was created to handle user input to control a given robot design.
basic colliders were given to objects to emulate normal rigidbody physics, but a bug to work out is the coral station walls having rotated hitboxes.