Created a more efficient and universal input system for multiple robots.
I created a more clean setup for using multiple different robots. The only thing required to do is to change which robot's scriptable object you put in the game manager and it puts the correct robot in with the correct controls.
PlayerController.cs takes the Scriptable Object and collects the PlayerInput from it and the InputInstructions script along with it to set up the proper keybinds for the according robot. Each robot sets up the drive controls and all additional inputs that that specific robot requires.