Implemented Finite State Machine
Implemented a finite state machine for the robots to more easily implement functionality. The reason was to not have any states/actions rely on each other, so as long as each state starts and ends properly, there shouldn't be any conflict. Implemented a compatable version, but only used KitBot as a demonstration. MukwonagoBot code was commented out and needs to be fixed.