Implemented MukwonagoBot Mechanism Movements
Implemented scripts for each subsystem for the climber, elevator, shoulder, elbow, and wrist. Additionally updated the state machine to include the idle state and intake state positions to make the mechanisms go to the correct position. Still need to figure out the elevator movement, and deal with actions such as scoring where there are four different levels, but not enough buttons.