Implemented Physics With Elevator
All components on the MukwonagoBot robot all have physics layers, so now the robot doesn't have to wory about self intersection. A Coral Zone was added to try and make coral near the coral manipulator apply a force to stay in the manipulator. The Elevator also moves using it's Rigidybody instead of transform with Vector3.lerp