Added Elevator Offset For Collisions And Cleaned Up Code and Physics
Added a 1 cm collision offset on the elevator to make sure it doesn't look like it is phasing into objects. Also adjusted the RobotInfo scriptable object to include initial tranform position information, and gave that to BaseRobot through an accessor. The Base Robot now automatically handles locking the robot vertically along the y axis and rotationally along the x and z axes.