Implemented smarter Coral Station drop logic
The coral station will only now drop Coral only for MukwonagoBot only when it is in the Intaking state. It will also allow you to start intaking and drop a Coral on the press. It additionally won't drop any more coral if you have a coral already. MukwonagoBot also now has a static variable to keep track of whether or not it is currenlyt holding a coral. Minor changes were made to the bumpers to reduce vertex count, Kitbot is having issues with the coralstation and wall collisoins.