Added an algae zone, fixed score reset bug, and closed unused scripts in IDE
Added an AlgaeIntakeZone with its own script onto the end of the elbow. Fixed a bug where when you reloaded the game after entering the menu scene, the score wouldn't reset to zero due to score being static. Also closed a bunch of unnecesary scripts in the IDE for better organization