Updated State Change Logic
Moved logic from handling states from the states themselves for the MukwonagoBotManager. When a button was pressed, it would immediately change the state, but I was telling the state to not do anything if a certain condition was true. Instead, I changed it so when a button is pressed, it will check that a certain condition isn't true before changing the state.