Added Spike Mark Coral and Algae
Added Coral and Algae onto the three Spike marks onto each of the alliances. It is very picky because the Coral uses a capsule collider, so the Algae is balancing on top of a rounded shape. That rounded shape is also somehow balancing upright on the ground with an Algae on top. Just don't touch it and it will be fine.