Implemented cycling scoring location logic
Added a button on left bumper to allow for the target algae intake and coral score locations to be cycled through. Additionally adjusted each of the states to go to different positions and confirmed positions based on the current Enum values. Additionally fixed  the logic that deals with confirm scoring, where if you let go of score and then confirm score, it would go back up to the score position with no buttons held down.