Implemented a UI script that highlights preset UI items
Created the UI script to highlight the correct buttons on the UI to indicate current preset values. highlights the current value to green and the next value in yellow. Runs efficiently by adding a subscriber to the static MukwonagoBotPresets.cs class to run exactly once when the value changes, and avoids making an instance of itself there.