Implemented a ScorableLocation class
Introduced the ScorableLocation class to act as a baseline for all locations game pieces can be scored. Made sure each location has access to a static Score Manager, alliance color, and target tag to look for. All implementations of this class are to extend this class and are meant to override the functionality of OnScored and IsValidScoringObject, and utilize those in thier own context. The ReefPole class was restructured to follow these instructions and currently still works.