Implemented Net scoring using ScorableLocation superclass
Created the Net.cs class to handle net scoring. First new implementation of the new ScorableLocation class and worked well.  Script handles scoring by remembering Algae in a HashSet and adds and removes through through ontriggerenter, stay, and exit. onunscored was included to deal with Algae falling out of the reef. attachted to an empty with a hitbox to implement.