Restructured the Field CAD and Implemented New models
Took models straight from onshape and imported them into blender, and with a couple settings tweaked, I put it straight into Unity with a couple on onshape measurements. the organization and heirarchy of the models is changed to be more consistent and interactable for later changes. Current field element prefabs for the cages and reef poles are currently left in and broken so I can use them as reference to add them back in.