Implemented Coral Station Logic with HumanPlayerStation superclass and namespace
Implemented a base superclass called HumanPlayerstation, which manages the general functionality of a human player station. this assumes that a station will have a box collider, and some logic to drop/spawn a game piece. leaves an abstract method to make the subclass fill in the logic. The superclass has built in editor logic to organize the editor so you just have to edit one script. Values are automatically updated and spawns a prefab of the scoring element and moves it around to show the spawn offset, and gets removed when the game starts. all extra code if in a preprocessor directive to not compile this code in the actual game.