Implemented updated UIs
Moved components such as Timer and ScoreManger to the GameManager object. Added data bindings to the uid menu from a new scriptable object. Scoremanager takes an instance of the scriptable object and just changes the values there. Same logic is applied to the Timer class, and it also additionally added delegate callback events to broadcast methods to other scripts about when the stages of the game change.