Implemented BaseZone and BargeZone class
Refactored the Starting Line class to use the zonebase class, which acts as a template for zones that award points based on robots being in or out of a zone on a given period of the match. cleaned up the starting line file and duplicated it to create the barge zone. A minor change was to cache robots that entered the zone before endgame started to not skip over them.