Refactored the robot functionality.
Heavily adjusted the base robot to be very modular. the BaseRobot now just handles the robot as a whole, and you can slap on subsystems to make a coherent robot. PoseLibrary and RobotPose were created to store poses and such information about a robot's poses set positions/rotations to scale with new robots and changes.