Implemented tunable settings
Removed normalization of vectors, as this is a simulation and it doesn't really need to be perfect realistic, and because it was causing issues with drive speeds. Added in drive and rotation force multipliers and an active breaking torque value that are tunable to give the swerve drive better values.